git_ci_gym/server/requirements.txt
openenv-core
fastapi
uvicorn
pydantic>=2.0
fastmcp
pytest
